@@ -1,9 +1,7 @@
 class HomeController < ApplicationController
-  layout 'ontology'
+  layout :determine_layout
 
   #RI_OPTIONS = {:apikey => $API_KEY, :resource_index_location => "#{$REST_URL}/resource_index/", :limit => 10, :mode => :intersection}
-
-  NOTES_RECENT_MAX = 5
 
   def index
     @ontologies_views = LinkedData::Client::Models::Ontology.all(include_views: true)
@@ -38,7 +36,10 @@ class HomeController < ApplicationController
     if $DISPLAY_RECENT.nil? || $DISPLAY_RECENT == true
       @recent_mappings = get_recent_mappings  # application_controller
     end
-    # calculate bioportal summary statistics
+    
+    organize_groups
+
+    # Calculate BioPortal summary statistics
     @ont_count = @ontologies.length
     @cls_count = LinkedData::Client::Models::Metrics.all.map {|m| m.classes.to_i}.sum
     @individuals_count = LinkedData::Client::Models::Metrics.all.map {|m| m.individuals.to_i}.sum
@@ -57,6 +58,16 @@ class HomeController < ApplicationController
       @direct_expanded_annotations = @ri_stats[:total][:ancestors]
     end
     @analytics = LinkedData::Client::Analytics.last_month
+
+    @ontology_names = @ontologies.map{ |ont| ["#{ont.name} (#{ont.acronym})", ont.acronym] }
+
+    @anal_ont_names = {}
+    @anal_ont_numbers = []
+    @analytics.onts[0..4].each do |visits|
+      ont = @ontologies_hash[visits[:ont].to_s]
+      @anal_ont_names[ont.acronym] = ont.name
+      @anal_ont_numbers << visits[:views]
+    end
   end
 
   def render_layout_partial
@@ -112,22 +123,21 @@ class HomeController < ApplicationController
     if params[:comment].nil? || params[:comment].empty?
       @errors << "Please include your comment"
     end
-    # verify_recaptcha is a method provided by the recaptcha plugin, returns true or false.
-    if ENV['USE_RECAPTCHA'] == 'true' && !session[:user]
+    if using_captcha? && !session[:user]
       if !verify_recaptcha
         @errors << "Please fill in the proper text from the supplied image"
       end
     end
 
     unless @errors.empty?
-      render :layout => feedback_layout
+      render layout: feedback_layout
       return
     end
 
     Notifier.feedback(params[:name],params[:email],params[:comment],params[:location]).deliver_now
 
     if params[:pop].eql?("true")
-      render :action => "feedback_complete", :layout => "popup"
+      render "feedback_complete", layout: "popup"
     else
       flash[:notice]="Feedback has been sent"
       redirect_to_home
@@ -136,22 +146,6 @@ class HomeController < ApplicationController
 
   def user_intention_survey
     render :partial => "user_intention_survey", :layout => false
-  end
-
-  def robots
-    if @subdomain_filter[:active]
-      robots = <<-EOF.gsub(/^\s+/, "")
-        User-agent: *\n
-        Disallow: /
-      EOF
-      render :text => robots, :content_type => 'text/plain'
-    else
-      robots = <<-EOF.gsub(/^\s+/, "")
-        User-Agent: *
-        Disallow:
-      EOF
-      render :text => robots, :content_type => 'text/plain'
-    end
   end
 
   def site_config
@@ -176,7 +170,7 @@ class HomeController < ApplicationController
     projects = LinkedData::Client::Models::Project.all;
     @user_projects = projects.select {|p| p.creator.include? @user.id }
 
-    render :partial => "users/details", :layout => "ontology"
+    render partial: "users/details", layout: determine_layout()
   end
 
   def feedback_complete
@@ -188,6 +182,21 @@ class HomeController < ApplicationController
   def validate_ontology_file
     response = LinkedData::Client::HTTP.post("/validate_ontology_file", ontology_file: params[:ontology_file])
     @process_id = response.process_id
+  end
+
+  private
+
+  # Dr. Musen wants 5 specific groups to appear first, sorted by order of importance.
+  # Ordering is documented in GitHub: https://github.com/ncbo/bioportal_web_ui/issues/15.
+  # All other groups come after, with agriculture in the last position.
+  def organize_groups
+    # Reference: https://lildude.co.uk/sort-an-array-of-strings-by-severity
+    acronyms = ["UMLS", "OBO_Foundry", "WHO-FIC", "CTSA", "caBIG"]
+    size = @groups.size
+    @groups.sort_by! { |g| acronyms.find_index(g.acronym[/(UMLS|OBO_Foundry|WHO-FIC|CTSA|caBIG)/]) || size }
+
+    others, agriculture = @groups.partition { |g| g.acronym != "CGIAR" }
+    @groups = others + agriculture
   end
 
 end
