@@ -1,4 +1,4 @@
-class CreateUsers < ActiveRecord::Migration
+class CreateUsers < ActiveRecord::Migration[4.2]
   def self.up
     create_table :users do |t|
       t.string :first_name, :last_name, :email, :phone, :user_name,:hashed_password
