@@ -1,4 +1,4 @@
-class CreateRatingTypes < ActiveRecord::Migration
+class CreateRatingTypes < ActiveRecord::Migration[4.2]
   def self.up
     create_table :rating_types do |t|
       t.string :name
