@@ -1,4 +1,4 @@
-class CreateEventItems < ActiveRecord::Migration
+class CreateEventItems < ActiveRecord::Migration[4.2]
   def self.up
     create_table "event_items", :force => true do |t|
       t.string   "event_type",    :limit => 50
