@@ -1,4 +1,4 @@
-class CreateVirtualApplianceUsers < ActiveRecord::Migration
+class CreateVirtualApplianceUsers < ActiveRecord::Migration[4.2]
   def self.up
     create_table :virtual_appliance_users do |t|
       t.integer :user_id
