@@ -1,4 +1,4 @@
-class ChangeVirtualApplianceUserIdType < ActiveRecord::Migration
+class ChangeVirtualApplianceUserIdType < ActiveRecord::Migration[4.2]
   def self.up
     change_column :virtual_appliance_users, :user_id, :string
   end
