@@ -31,8 +31,9 @@ function set_last_params(params) {
   //console.log(bp_last_params);
 }
 
-function insertSampleText() {
+function insertSampleText(event) {
   "use strict";
+  event.preventDefault();
   var text = "Melanoma is a malignant tumor of melanocytes which are found predominantly in skin but also in the bowel and the eye.";
   jQuery("#annotation_text").focus();
   jQuery("#annotation_text").val(text);
@@ -46,7 +47,7 @@ function get_annotations() {
   jQuery("#annotator_error").html("");
 
   // Validation
-  if (jQuery("#annotation_text").val() === jQuery("#annotation_text").attr("title")) {
+  if (!jQuery("#annotation_text").val()) {
     jQuery("#annotator_error").html("Please enter text to annotate");
     return;
   }
@@ -161,8 +162,10 @@ function get_annotations() {
 var displayFilteredColumnNames = function() {
   "use strict";
   var column_names = [];
+  var header_text;
   jQuery(".bp_popup_list input:checked").closest("th").each(function() {
-    column_names.push(jQuery(this).attr("title"));
+    header_text = this.childNodes[0].textContent.trim();
+    column_names.push(header_text);
   });
   jQuery("#filter_names").html(column_names.join(", "));
   if (column_names.length > 0) {
@@ -560,6 +563,10 @@ jQuery(document).ready(function() {
   filter_match_type.init();
   filter_matched_ontologies.init();
   filter_matched_classes.init();
+
+  jQuery("#annotator-help").on("click", bpPopWindow);
+
+  jQuery("#annotations_container").hide();
 }); // doc ready
 
 
