@@ -1,7 +1,12 @@
 class SubmissionsController < ApplicationController
 
+<<<<<<< HEAD
   layout 'ontology'
   before_action :authorize_and_redirect, :only=>[:edit, :update, :create, :new]
+=======
+  layout :determine_layout
+  before_action :authorize_and_redirect, :only=>[:edit,:update,:create,:new]
+>>>>>>> fa32c2b620cdc8e626739a67cde7cac7ebdda1d5
 
   # When getting "Add submission" form to display
   def new
@@ -25,6 +30,7 @@ class SubmissionsController < ApplicationController
     # Make the contacts an array
     params[:submission][:contact] = params[:submission][:contact].values
 
+<<<<<<< HEAD
     # Get the submission metadata from the REST API
     json_metadata = JSON.parse(Net::HTTP.get(URI.parse("#{REST_URI}/submission_metadata?apikey=#{API_KEY}")))
     @metadata = json_metadata
@@ -62,7 +68,25 @@ class SubmissionsController < ApplicationController
         else
           redirect_to "/ontologies/success/#{@ontology.acronym}"
         end
+=======
+    @submission = LinkedData::Client::Models::OntologySubmission.new(values: submission_params)
+    @ontology = LinkedData::Client::Models::Ontology.get(params[:submission][:ontology])
+    
+    # Update summaryOnly on ontology object
+    @ontology.summaryOnly = @submission.isRemote.eql?("3")
+    @ontology.update
+    
+    @submission_saved = @submission.save
+    if !@submission_saved || @submission_saved.errors
+      @errors = response_errors(@submission_saved) # see application_controller::response_errors
+
+      if @errors[:error][:uploadFilePath]
+        @errors = ["Please specify the location of your ontology"]
+      elsif @errors[:error][:contact]
+        @errors = ["Please enter a contact"]
+>>>>>>> fa32c2b620cdc8e626739a67cde7cac7ebdda1d5
       end
+
       render "new"
     else
       redirect_to "/ontologies/success/#{@ontology.acronym}"
@@ -93,7 +117,10 @@ class SubmissionsController < ApplicationController
     # Make the contacts an array
     params[:submission][:contact] = params[:submission][:contact].values if !params[:submission][:contact].nil?
 
+    params[:submission][:contact].delete_if { |c| c[:name].empty? || c[:email].empty? }
+
     @ontology = LinkedData::Client::Models::Ontology.get(params[:submission][:ontology])
+<<<<<<< HEAD
 
     #submissions = @ontology.explore.submissions
     #@submission = submissions.select {|o| o.submissionId == params["id"].to_i}.first
@@ -121,10 +148,16 @@ class SubmissionsController < ApplicationController
     end
 
     @submission.update_from_params(params[:submission])
+=======
+    submissions = @ontology.explore.submissions
+    @submission = submissions.select {|o| o.submissionId == params["id"].to_i}.first
+>>>>>>> fa32c2b620cdc8e626739a67cde7cac7ebdda1d5
 
+    @submission.update_from_params(submission_params)
     # Update summaryOnly on ontology object
     @ontology.summaryOnly = @submission.isRemote.eql?("3")
-    @ontology.save
+    @ontology.update
+    error_response = @submission.update
 
     error_response = @submission.update(cache_refresh_all: false)
 
@@ -133,6 +166,17 @@ class SubmissionsController < ApplicationController
     else
       redirect_to "/ontologies/#{@ontology.acronym}"
     end
+  end
+
+  private
+
+  def submission_params
+    p = params.require(:submission).permit(:ontology, :description, :hasOntologyLanguage, :prefLabelProperty,
+                                           :synonymProperty, :definitionProperty, :authorProperty, :obsoleteProperty,
+                                           :obsoleteParent, :version, :status, :released, :isRemote, :pullLocation,
+                                           :filePath, { contact:[:name, :email] }, :homepage, :documentation,
+                                           :publication)
+    p.to_h
   end
 
 end
