@@ -1,11 +1,10 @@
-# config valid only for Capistrano 3.1
-lock '3.4.1'
+#lock '3.7.1'
 
 set :application, 'bioportal_web_ui'
 
-set :repo_url, "git@github.com:ncbo/#{fetch(:application)}.git"
+set :repo_url, "https://github.com/ncbo/#{fetch(:application)}.git"
 
-#set :deploy_via, :remote_cache
+set :deploy_via, :remote_cache
 
 # Default branch is :master
 # ask :branch, proc { `git rev-parse --abbrev-ref HEAD`.chomp }
@@ -26,7 +25,7 @@ set :deploy_to, "/srv/rails/#{fetch(:application)}"
 # set :pty, true
 
 # Default value for :linked_files is []
-#set :linked_files, %w{config/bioportal_config.rb config/database.yml public/robots.txt config/newrelic.yml}
+#set :linked_files, %w{config/bioportal_config.rb config/database.yml public/robots.txt}
 
 # Default value for linked_dirs is []
 #set :linked_dirs, %w{bin log tmp/pids tmp/cache public/system public/assets config/locales}
@@ -49,9 +48,17 @@ set :passenger_restart_with_touch, true
 
 namespace :deploy do
 
+  desc 'display remote system env vars' 
+  task :show_remote_env do
+    on roles(:all) do
+      remote_env = capture("env")
+      puts remote_env
+    end
+  end
+
   desc 'Incorporate the bioportal_conf private repository content'
-  #Get cofiguration from repo if PRIVATE_CONFIG_REPO is set 
-  #(should be set to NCBO private config repo)
+  #Get cofiguration from repo if PRIVATE_CONFIG_REPO env var is set 
+  #or get config from local directory if LOCAL_CONFIG_PATH env var is set 
   task :get_config do
      if defined?(PRIVATE_CONFIG_REPO)
        TMP_CONFIG_PATH = "/tmp/#{SecureRandom.hex(15)}"
@@ -59,8 +66,12 @@ namespace :deploy do
           execute "git clone -q #{PRIVATE_CONFIG_REPO} #{TMP_CONFIG_PATH}"
           execute "rsync -av #{TMP_CONFIG_PATH}/#{fetch(:application)}/ #{release_path}/"
           execute "rm -rf #{TMP_CONFIG_PATH}"
-      end
-    end
+       end
+     elsif defined?(LOCAL_CONFIG_PATH) 
+       on roles(:web) do
+          execute "rsync -av #{LOCAL_CONFIG_PATH}/#{fetch(:application)}/ #{release_path}/"
+       end
+     end
   end
 
   desc 'Restart application'
@@ -71,7 +82,6 @@ namespace :deploy do
     end
   end
 
-  #before :started, :get_config
   after :updating, :get_config
   after :publishing, :restart
 
