@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -13,27 +12,27 @@
 
 ActiveRecord::Schema.define(version: 20150806211835) do
 
-  create_table "analytics", force: :cascade do |t|
-    t.string   "segment",    limit: 255
-    t.string   "action",     limit: 255
-    t.string   "bp_slice",   limit: 255
-    t.string   "ip",         limit: 255
-    t.integer  "user",       limit: 4
-    t.text     "params",     limit: 65535
+  create_table "analytics", id: :integer, force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+    t.string "segment"
+    t.string "action"
+    t.string "bp_slice"
+    t.string "ip"
+    t.integer "user"
+    t.text "params"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "timeouts", force: :cascade do |t|
-    t.string   "path",        limit: 255
-    t.integer  "ontology_id", limit: 4
-    t.text     "concept_id",  limit: 65535
-    t.text     "params",      limit: 65535
-    t.datetime "created"
+  create_table "timeouts", id: :integer, force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+    t.string "path"
+    t.integer "ontology_id"
+    t.text "concept_id"
+    t.text "params"
+    t.timestamp "created"
   end
 
-  create_table "virtual_appliance_users", force: :cascade do |t|
-    t.string   "user_id",    limit: 255
+  create_table "virtual_appliance_users", id: :integer, force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+    t.string "user_id"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
